+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * % (MODULO) operator (9%7 == 2)
 * Fix 'undefined' column if alias is not exists
 * x > ALL (subquery), x > SOME/ANY (subquery)
+* Check if table exists in the database
+* Multiple user defined functions arguments, like TRIPLE(a,b,c)
 
 ### 0.0.13 (06.11.2014)